refactor: rename components, change component imports, change styles
