+++ src/components/index.ts
@@ -1,3 +1,5 @@
 export { Search } from './search/search.component';
-export { Cards } from './cards/cards.component.tsx';
-export { ErrorBoundary } from './error/error.tsx';
+export { Cards } from './cards/cards.component';
+export { ErrorBoundary } from './error/error.component';
+export { NotFound } from './not-found/not-found.component';
+export { Loader } from './loader/loader.component';
