+++ src/index.css
@@ -7,6 +7,7 @@
     rgba(0, 187, 191, 1) 0%,
     rgba(0, 67, 87, 1) 100%
   );
+  min-height: 100vh;
 }
 
 .app {
